Split the code up into logical modules
